Level and Difficulty Changes
Added a system that restarts the room but increments the balloon speed.
At a certain level the level will reset at 1.
Added a collider to destroy balloons if they get too high since the ballons travel higher with a  15 second lifespan if they are moving faster.

Lifespan is not really needed anymore, however, I left it just in case.