Created popping animation
Found popping animation on asset store and added to to the balloons. However a problem exists that the balloon gets destroyed before the animation has time to play so the user doesnt see it. I'm going to have to rework the DecrementInteractables to work a around that. I'm just going to implement "particle.isPlaying" to know if it should destroy it.